PartSolve: Simplify contract between intermediate and solvers
Prior to this commit, the `PartSolve` trait required implementers to:

- Return (assuming success) a `Box<dyn Any>` from the `parse` method.

- Grant both the `part_one` and `part_two` methods a (read-only)
  &Box<dyn Any>.

There were two main issues with this contract.

- Box<T> is a *pointer* to a uniquely-owned heap allocation of type T.
  So by asking for an &Box<dyn Any> in the part functions, we were in
  effect asking for a pointer to a pointer to a heap allocation. This is
  needless indirection, and makes the code harder to read.

- This also created an awkward guarantee. No matter what, the driver
  code (in main.rs) is required to have a Box<dyn Any> to give out
  references to. This is actually an implementation detail that we were
  leaking into the trait contract, and meant that the driver code and
  the trait were tightly coupled. (Hence I had to go update all these
  solvers.)

Now, `PartSolve` still requires the `parse` method to return a
`Box<dyn Any>`, but it now only requires that `part_one` and `part_two`
accept a `&dyn Any`. This is a reference to a trait object, which is
a more ergonomic way to work with dynamically-typed values in Rust. It's
also a single pointer, like the `Box`, with a defined lifetime that is
easier to infer from.

I did have to go do shotgun surgery, but I don't think I will again with
this approach.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>