Avoid using `std` paths where `core` suffices
To make it more obvious where dependencies exist on `std`, we can use
the `std_instead_of_core` lint from Clippy to deny usage of `std` paths
where `core` paths exist to the exact same item. To this end, this
commit upgrades this lint (which is allow-by-default) to `deny` and
fixes known violations.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>