Fix broken intra-doc link to `Any::downcast_ref`
The `downcast_ref` method is not defined on the `Any` trait itself but
rather on the `dyn Any`, `dyn Any + Send`, and `dyn Any + Send + Sync`
types. Thus, a different path was needed to make the link work. For the
sake of the reader, we use the same display text as before and hide this
complexity in the link target.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>